samples: video: capture: Fix camera fixture test
The logs that the fixture test are based on are not printed anymore due
to recent changes from "printk" to "LOG_DBG". Change the log level so
that it can work again.

Remove "Capture started" as it is not a relevant test regex. Logs are
sometimes not correctly printed in the console, for example, we can see:

DEBUG   - DEVICE: muart:~$ [mapture started

As "Capture started" is a one-time log, fixture test will then fail due
to timeout.

Also, consolidate other regex.

Signed-off-by: Phi Bang Nguyen <phibang.nguyen@nxp.com>